Github Push & Testing Connection
Restarting this. The project that I made for phase 1 and 2 got messed up cuz I was trying to delete some assets off.

Then. I ended up deleting some parts that shoud *not* have been touched. It kind of became a mess after that.

But that's besides the point, anyways. I'm recreating this [again] and turning it in. Hooooooo